Refactor /user/{x}/vote/{x}/
Reuses the result from /user/{x}/vote/ to avoid repeating the userExists
code.

Adds a test that the output error message is the expected message, when
a 404 occurs.

Also removed a stray ref to System.Net.Http.Formatting (again), as we
have 2 references to the lib. One to 5.2.0.0, one to 5.2.2.0.