Factor out email sending to another class
Creates the MailSender class to send an email from the voting.app@gmail.com
address to a provided email with the provided subject and message.

We also go back to using concatenated strings instead of the verbatim
string, because the verbatim string messes up the indentation by adding
whitespace at the beginning of indented lines.