Remove height warning in console
The warning was that we were trying to draw with no space to draw
(because the chart was hidden). Now, we just don't draw if we are already
hiding the chart.

Also adds on the "Dataloss from migrations" flag, as this was coming up
a lot when changing between branches