Fix for axis ordering with MarkerSeries
The data being passed in to the MarkerSeries was an array of strings,
while the columns used an array of objects with Name properties. Here
we make the MarkerSeries use the same array of objects.

We can now promote the MarkerSeries to being drawn before the ColumnSeries
so it is consistently series-0 in the CSS. This means the legend item is
styled correctly too