Fix for tabbing issue on Create page
By having a href on the Strategy [?] info box, we could tab to it but
not to any other tooltip.

The link wasn't really required, so we remove it.

Also straightens out the div types from a previous merge conflict which
had left the Strategy type in a section of its own.