Default to compiling in Release mode
By default, if no configuration was specified, the app would build in
Debug mode. This meant that the Min files were not compiled, so the
production site had issues with them.

We also remove the JSDebug script, in favour of only using Min files if
we have attached a debugger. This will allow us to debug things on prod
if necessary, without causing the full files to be used when not required