Use release mode for live sites
When building locally, we can select to be in Debug mode.
When deploying to the live site, it will use this setting.
Previously, it was using Debug and therefore not running the Minification
scripts, so couldn't find /Scripts/Min/xyz.js

The Dev site avoided this because it does not use the Minified scripts.