Fill in the blanks
All buttons now have a general page for their setting and can return to
the main poll page.
The manage page puts the gva-unlockable-option into use to replace some
of the other hard-coded versions, and renames a few headings.

We will probably want to pull the expiration date logic out to a
different page soon.