Add initial password reset mechanism
Email service is disabled in the code, need to introduce an interface
to make use of IOC container.  Email should be executable on dev
machines as well as dev server.

The intention is for email confirmation to be added. A nice to have
would be to use MVC render engine to contain html templates for emails.