Remove email sending from poll creation
When creating a poll, you're either anonymous, in which case we don't
have an email, so we can't send one, or you're logged in, so you'll see
the poll etc. on the dashboard. Thus the attempt to send an email on
creation is redundant (and will actually crash as it's always assuming
there's an email address!).