Move to results when poll closes
Merges the PollHeading directive and PollHeadingController into a single
file, as we need to use the isolated scope in the directive and it
doesn't like sharing with the controller.

We provide an optional parameter for the PollHeading to call a callback
method if/when a poll expires. If the poll heading updates to close the
poll or the poll has already closed before the page load, the user is
redirected to the results

Because the results page doesn't provide this callback, it is only called
if the user lands on the voting page after the expiry.

This callback is not executed if the poll never expires.