Display username in place of email
If the user has already cast a vote with a name, display them with that
name rather than the email address.

This is to help with open polls, where a user can join the poll and gain
a token without having been explicitly invited. Beforehand, this would
display them with a blank email.