Fix saving issue in open polls
Some of the voters had a null email, due to voting without an invitation.
This meant that asking for t.Email.Equals() would try calling Equals on
null, which falls over.

We now fall back on the voter name if there is no email. This is not
unique, but we have nothing else identifiable about the user that we can
use.