Do not delete voters
This was an issue of timing; the poll creator would be on a page where
the poll.Voters had already been fetched. When a new user came and voted,
their name would be added to the server's poll.Voters, but not updated
in the poll creator's local copy. The poll creator would then submit the
updated poll, with an outdated version of poll.Voters. The backend would
identify that the new voter was missing from this submission, and assume
their token had been invalidated.

The situation will not occur with anybody who was invited by email, as the
poll creator has to save their addition in order for them to be able to
vote in the first place.