JSLint - Disallow empty blocks
We only had the one empty block with "if (!token) { // Tell the user }".
I noticed the submitVote code was almost identical in all of the classes,
with the exception of forming the vote array.

Because of this, I have pulled the code up to the VotingPageController
level, and let each vote type controller form the votes array themselves.