Tidy up voting refactor
We only want to submit a single vote, not multiple votes. All of the
cshtml files bind to submitVote too, so we change our own function
instead.

The resultsLink now lives in the file doing the redirect, so there is no
need to add it to the scope.

The getVotes method lives in the calling scope, not the local scope, so
we have to use "this.getVotes" in order to call it.