JSLint - Avoid "this"
"use strict" was flagging up that calling "this.*" was dodgy.
We want to pass up a function from a child scope, without knowing which
strategy the child is using.

We now set the getVotes method to return [] by default, and add a method
to the scope to override this from the child strategy.

Also rewinds the removal of resultsLink from the scope, as this is used
by the "Go to Results" button!