Display correct invitees count on Manage page
The Invitees only includes people we have sent emails to, as this is all
that we display when we go to the Invitees page. This is sometimes less
than the total count of ballots, as some people may join the poll through
the link without having been invited by email.

The count of invitees needs to reflect the number of email addresses,
not the number of ballots.