Save on "Invite Pending"
Revives the /manage/xyz/invitation route, and uses it to send invitations
separately from saving the poll.

The "POST /manage/xyz/invitation" now takes a list of emails to invite,
as some of these may not yet be saved to the poll, and we don't want to
call "Save" beforehand because we may still want to discard some deletions.

As a side effect, it is now possible to re-invite emails from the backend
although the front-end doesn't yet support this.