First pass at new Options look
Uses the trash symbol as a "Delete" and the ban symbol as a "Cancel",
following the research amongst the office.

Adds a note that "No options have been added"

Groups together items to be displayed together when edit mode is on/off

Consolidates the add-option method do do the clearing of the field. Also
removes the unused description toggle.