Handle all cases via the backend
Tests all possible configurations of new/pending/existing invites against
the save/invite/delete actions we expect to run, and checks that
everything is consistent.

We gut the ManageController's handling of invitations and move everything
into the ManageInvitationController.

We also split apart the ManagePollBallotRequestModel to have a
ManageInvitationRequestModel for POST requests, and a
ManageInvitationResponseModel for the GET requests.