Handle sending invitations
We want to mark any pending invitations as requiring something to be sent
and none of the others. It doesn't matter if we have duplicates in the
Invitations and pendingUsers, as they will share the same ManageGuid so
will be treated as a single ballot.

This will still preserve any deleted items, which are either applied
when saved or discarded when cancelled.