Remove TokenService from BasicVoteController
VotingPageController makes a call to get the voter token, so store this on
the scope, instead of getting it in the child scope. It's cached, so it's
not an api call, but we already have this information in scope, and
removing it makes it a bit easier to remove the VoteService from the
BasicVoteController.