Always use Poll ID, not ManageId
Pulls out some of the logic searching the database for a corresponding
UUID given an ManageId

We want to see results such as "A change of a poll setting on management
caused errors to be thrown on the voting side", which is harder to do
if we're using two different IDs.

Now, any changes to management properties are logged against the PollId,
not the ManageId.