Metrics for adding and removing votes
When we cast a vote, add a metric.
When votes are cleared out for any reason (deleting related options/ballots,
changing poll type, or explicitly deleting the vote), add a metric

As a side note, some of these tests have a ridiculous amount of setup
just to do something simple. Building up the request to delete a ballot in
particular appears to have a great deal of redundancy.