Create the vote metrics properly
Fills in the fields on the metric and stores them.
I have also repurposed the Value field, as we were previously only using
this for Status Codes, but almost always setting it to "OK". Status codes
make sense for Errors and Page requests, but not so much for
already-succeeded requests.

Instead, we use it to indicate the options in the poll which the votes
correspond to, allowing us to group them more easily and track over time