Force logout if token is invalid
If the call to get the polls for the registered dashboard fails then
remove the local auth token. The observer function will cause the homepage
to switch the dashboard over to the unregistered dashboard.

Trying to try this all together to show the login dialog if it fails just
wouldn't work.

This fixes VA-391 and VA-406.