Fix issues with migrations
Migration OptionToChoice was created at the same time as
LastUpdated_CreatedDate_Utc, both in different branches. As a result,
LastUpdated_CreatedDate_Utc has no knowledge of the model changes in
OptionToChoice. This is causing EF to try to scaffold the changes in
OptionToChoice all over again.

To fix this, we need to remove LastUpdated_CreatedDate_Utc from the 19th
and re-scaffold it. The script is the same, but the model in the migration
history is now the latest.

Before running the migration, you need to roll back to OptionToChoice.